+++ custom_components/danfoss_air_ccm/__init__.py
@@ -9,6 +9,7 @@ PLATFORMS = [
     "number",
     "switch",
     "button",
+    "select",
 ]