+++ custom_components/danfoss_air_ccm/switch.py
@@ -13,6 +13,8 @@ async def async_setup_entry(hass, entry, async_add_entities):
     async_add_entities(
         [
             DanfossBypassSwitch(coordinator),
+            DanfossBoostSwitch(coordinator),
+            DanfossBoostAutoSwitch(coordinator),
         ]
     )
 
@@ -34,3 +36,40 @@ class DanfossBypassSwitch(DanfossEntity, SwitchEntity):
 
     async def async_turn_off(self, **kwargs):
         await self.coordinator.set_bypass(False)
+
+class DanfossBoostSwitch(DanfossEntity, SwitchEntity):
+
+    _attr_name = "Boost"
+
+    def __init__(self, coordinator):
+        super().__init__(coordinator)
+        self._attr_unique_id = "danfoss_boost"
+
+    @property
+    def is_on(self):
+        return self.coordinator.data["boost"]
+
+    async def async_turn_on(self, **kwargs):
+        await self.coordinator.set_boost(True)
+
+    async def async_turn_off(self, **kwargs):
+        await self.coordinator.set_boost(False)
+
+
+class DanfossBoostAutoSwitch(DanfossEntity, SwitchEntity):
+
+    _attr_name = "Boost Auto"
+
+    def __init__(self, coordinator):
+        super().__init__(coordinator)
+        self._attr_unique_id = "danfoss_boost_auto"
+
+    @property
+    def is_on(self):
+        return self.coordinator.data["boost_auto"]
+
+    async def async_turn_on(self, **kwargs):
+        await self.coordinator.set_boost_auto(True)
+
+    async def async_turn_off(self, **kwargs):
+        await self.coordinator.set_boost_auto(False)